Issue #5: adapt to GitHub API break (switch from int to int64 in release id and release asset id fields)

+++ uploader/client.go
@@ -13,15 +13,15 @@ type Client interface {
 	GetReleaseByTag(tagName string) (Release, Response, error)
 	CreateRelease(release Release) (Release, Response, error)
 	UpdateRelease(release Release) (Release, Response, error)
-	DeleteRelease(releaseId int) (Response, error)
+	DeleteRelease(releaseId int64) (Response, error)
 	DeleteTag(tagName string) (Response, error)
-	ListReleaseAssets(releaseId int) ([]ReleaseAsset, Response, error)
-	DeleteReleaseAsset(assetId int) (Response, error)
-	UploadReleaseAsset(releaseId int, assetName string, assetFile *os.File) (ReleaseAsset, Response, error)
+	ListReleaseAssets(releaseId int64) ([]ReleaseAsset, Response, error)
+	DeleteReleaseAsset(assetId int64) (Response, error)
+	UploadReleaseAsset(releaseId int64, assetName string, assetFile *os.File) (ReleaseAsset, Response, error)
 }
 
 type Release interface {
-	GetID() int
+	GetID() int64
 	GetName() string
 	GetBody() string
 	SetBody(body string)
@@ -41,6 +41,6 @@ type Response interface {
 }
 
 type ReleaseAsset interface {
-	GetID() int
+	GetID() int64
 	GetName() string
 }
